linear: add perf model that fails to max
This model is the same as the standard model, but does not assume that
the |QI|=n will fit within the bounds of the emperical model. Instead,
if n is out of range, the max valid n is used. This is class is just
used for benchmarking where we want workloads bigger than the
performance model has measured values for.